pdftesting

+++ eSiroi.Resource/Entities/Uniowner.cs
@@ -19,7 +19,7 @@ namespace eSiroi.Resource.Entities
         [StringLength(30)]
         public string NewDagNo { get; set; }
 
-        public int ? ownno { get; set; }
+        public long ? ownno { get; set; }
 
         [StringLength(30)]
         public string NewPattaNo { get; set; }
